Move sets into their own tests #61
